chore(license): move back to MIT license
The Apache 2 license has a number of incompatibilities with the GPL, which is widely used by communities like WordPress. Keeping compatibility with them is a major priority so I don't see any reason to move to Apache 2 at this moment. We've done a little legal work and the reasons for moving don't seem compelling enough right now.